adding tests for parameter names obtained through reflection #1624
