+++ javaparser-symbol-solver-testing/src/test/java/com/github/javaparser/symbolsolver/reflectionmodel/ReflectionMethodDeclarationTest.java
@@ -31,48 +31,39 @@ import static org.junit.Assert.assertEquals;
 public class ReflectionMethodDeclarationTest {
 
     @Test
-    public void testGetSignature() {
+    public void testParameterNameOnClassesFromTheStdLibrary() {
         TypeSolver typeResolver = new ReflectionTypeSolver();
 
         ResolvedClassDeclaration object = new ReflectionClassDeclaration(Object.class, typeResolver);
         ResolvedInterfaceDeclaration list = new ReflectionInterfaceDeclaration(List.class, typeResolver);
 
-        ResolvedMethodDeclaration hashCode = object.getAllMethods().stream().filter(m -> m.getName().equals("hashCode")).findFirst().get().getDeclaration();
         ResolvedMethodDeclaration equals = object.getAllMethods().stream().filter(m -> m.getName().equals("equals")).findFirst().get().getDeclaration();
         ResolvedMethodDeclaration containsAll = list.getAllMethods().stream().filter(m -> m.getName().equals("containsAll")).findFirst().get().getDeclaration();
         ResolvedMethodDeclaration subList = list.getAllMethods().stream().filter(m -> m.getName().equals("subList")).findFirst().get().getDeclaration();
 
-        assertEquals("hashCode()", hashCode.getSignature());
-        assertEquals("equals(java.lang.Object)", equals.getSignature());
-        assertEquals("containsAll(java.util.Collection<? extends java.lang.Object>)", containsAll.getSignature());
-        assertEquals("subList(int, int)", subList.getSignature());
+        assertEquals("arg0", equals.getParam(0).getName());
+        assertEquals("arg0", containsAll.getParam(0).getName());
+        assertEquals("arg1", containsAll.getParam(1).getName());
+        assertEquals("arg0", subList.getParam(0).getName());
+        assertEquals("arg1", subList.getParam(1).getName());
+    }
+
+    class Foo {
+        void myMethod(int a, char c) {
+
+        }
     }
 
     @Test
-    public void testGetGenericReturnType() {
-        TypeSolver typeResolver = new ReflectionTypeSolver();
+    public void testParameterNameOnClassesFromThisProject() {
+        TypeSolver typeResolver = new ReflectionTypeSolver(false);
 
-        ResolvedInterfaceDeclaration map = new ReflectionInterfaceDeclaration(Map.class, typeResolver);
+        ResolvedClassDeclaration foo = new ReflectionClassDeclaration(Foo.class, typeResolver);
 
-        ResolvedMethodDeclaration put = map.getAllMethods().stream().filter(m -> m.getName().equals("put")).findFirst().get().getDeclaration();
-        assertEquals(true, put.getReturnType().isTypeVariable());
-        assertEquals(true, put.getReturnType().asTypeParameter().declaredOnType());
-        assertEquals("java.util.Map.V", put.getReturnType().asTypeParameter().getQualifiedName());
+        ResolvedMethodDeclaration myMethod = foo.getAllMethods().stream().filter(m -> m.getName().equals("myMethod")).findFirst().get().getDeclaration();
+
+        assertEquals("arg0", myMethod.getParam(0).getName());
+        assertEquals("arg1", myMethod.getParam(1).getName());
     }
 
-    @Test
-    public void testGetGenericParameters() {
-        TypeSolver typeResolver = new ReflectionTypeSolver();
-
-        ResolvedInterfaceDeclaration map = new ReflectionInterfaceDeclaration(Map.class, typeResolver);
-
-        ResolvedMethodDeclaration put = map.getAllMethods().stream().filter(m -> m.getName().equals("put")).findFirst().get().getDeclaration();
-        assertEquals(true, put.getParam(0).getType().isTypeVariable());
-        assertEquals(true, put.getParam(0).getType().asTypeParameter().declaredOnType());
-        assertEquals("java.util.Map.K", put.getParam(0).getType().asTypeParameter().getQualifiedName());
-
-        assertEquals(true, put.getParam(1).getType().isTypeVariable());
-        assertEquals(true, put.getParam(1).getType().asTypeParameter().declaredOnType());
-        assertEquals("java.util.Map.V", put.getParam(1).getType().asTypeParameter().getQualifiedName());
-    }
 }
